renaming of translation component; removing of C++11 features
