+++ components/CMakeLists.txt
@@ -62,8 +62,8 @@ add_component_dir (interpreter
     miscopcodes opcodes runtime scriptopcodes spatialopcodes types
     )
 
-add_component_dir (translation_data
-    translation_data
+add_component_dir (translation
+    translation
     )
 
 include_directories(${BULLET_INCLUDE_DIRS})
